+++ backend/packages/files/parsers/github.py
@@ -5,7 +5,7 @@ from langchain.document_loaders import GitLoader
 from langchain.schema import Document
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from models import Brain, File
-from utils.file import compute_sha1_from_content
+from packages.files.file import compute_sha1_from_content
 from utils.vectors import Neurons
